XojoKeywords() added to track what is a Keyword vs. Identifier
This allows us to handle parenthesis addition properly. For example,
prior to this change Format Code didn't know if an additional
keyword of "Abc" was a Keyword (i.e. should be followed by a space)
or just some other identifier. Now it does, so it knows "abc(1)" should
be written as Abc(1) instead of Abc (1).